added desktop changes
