+++ src/components/SignInButtons/AppleSignIn/index.android.js
